requirements.txt
requests
requests[socks]
pyyaml
rich
pydantic>=2.0
